✅ [Сборка 1210] Выпуск нового релиза - v0.3.2

❤ [Ядро] Исправления для Shell'a
❤ [Ядро] Исправления для виртуальной файловой системы (VFS)
❤ [Ядро] Исправления для функций для работы с видеобуфером
❤ [Ядро] Исправления для функций для работы с памятью
❤ [Ядро] Исправления для функций для работы с строками
❤ [Ядро] Поддержка PSF (Совместная работа с SynapseOS)
❤ [Ядро] Теперь все ядро использует PSF
❤ [Ядро] Обновлен экран смерти
❤ [Ядро] Обновлен экран переполнение памяти буфера
❤ [Ядро] Вырезана старая методика работы шрифтов (Sayori Font File)
❤ [Ядро] Зарегистрированы новые обработчики прерывания 
❤ [Ядро] Новая методика поиска файлов VFS. Теперь путь монтирования
имеет вес.
❤ [Ядро] Добавление аргумента загрузки nocoms (Отключает логирование в
COM's)
❤ [Ядро] Убрана инициализация при запуске всех портов COM
❤ [Ядро] Возможность выключения тестирования AC97 & PC-Speaker
❤ [Ядро] Исправление утечки setSysPath
❤ [Ядро] Исправление утечки setUserName
❤ [Ядро] Исправление утечки setHostname
❤ [Ядро] В shell возращена функция view
❤ [Ядро] Сортировка команд
❤ [Ядро] Улучшение view + поддержка относительного пути
❤ [Ядро] Добавлена поддержка Machinist API
❤ [Ядро] Поддержка языка программирования C++
❤ [Ядро] Поддержка virt2phys
💫 [Драйвера] Добавление структуры APIC
💫 [Драйвера] Начало разработки NatSuki [Milla]
💫 [Драйвера] Поддержка float (чисел с плавающей точкой)
💫 [Драйвера] Полноценная поддержка PCI (Совместная работа с SynapseOS)
💫 [Драйвера] Поддержка драйвера AC97 (только QEMU)
💫 [Драйвера] Добавлена структура для WAV-файла
💫 [Драйвера] Добавлен драйвер для мыши
💫 [Драйвера] [PCI] Теперь отображается ID Поставщика + Запись в ЛОГ
💫 [Драйвера] [PCI] Добавление поставщиков Apple + Red Hat + AMD (0x1022)
💫 [Драйвера] Поддержка ATA в режиме PIO
💫 [Драйвера] Отображение емкости HDD при подключении через ATA
💫 [Драйвера] LucarioFS (в режиме только чтения)
📂 [Файлы] Поддержка относительного пути
📦 [ПО] Встраивание Parallel Desktop в SayoriOS (автор @NDRAEY)
📦 [ПО] Обновление meminfo
📦 [ПО] Исправление для ELF
📦 [ПО] Исправление CPUInfo для процессоров Intel
📦 [ПО] Добавлено определение в CPUInfo для VMWare (не протестировано)
📦 [ПО] Добавлен курсор и отображение FPS в GUI
📦 [ПО] И уборка мусора после запуска GUI
📦 [ПО] Для BootScreen можно теперь отключить логирование
📦 [ПО] Возращена мелодия в shell (через PC-Speaker)
📦 [ПО] Обновлена методика обработки команд Shell'a
📦 [ПО] Добавлено Пианино (игра при помощи PC-Speaker'a)
🔌 [Прочие] Изменения в Grub > Отключение NatSuki
🔌 [Прочие] Поправка GRUB загрузчика
🔌 [Прочие] Изменения в Grub > Отключение NatSuki

Спасибо вам за то, что остаетесь с нами!
Обновление ядра приготовили для вас: Команда разработки SayoriOS Team
[­Никита Пиминов и Drew Pavlenko ]

Наша команда напоминает, что ядро является нестабильным, если вы решите
запустить его на реальном железе, то ВЫ и только ВЫ в ответе за ваше
сломанное оборудование.

+++ kernel/include/fs/sefs.h
@@ -2,16 +2,15 @@
 /**
  * @brief Структура кол-ва файлов
  */
-typedef struct{
+typedef struct {
    uint32_t nfiles; // Число файлов в ramdisk.
 } sefs_header_t;
 
 /**
  * @brief Структура файлов
  */
-typedef struct
-{
-   uint8_t index;             ///< Индекс
+typedef struct {
+   uint32_t index;             ///< Индекс
    uint32_t magic;            ///< Магическое число для проверки ошибок.
    char name[128];            ///< Имя файла
    uint32_t offset;           ///< Смещение в sefs, указывающее откуда начинается файл.
